qq-bot-py/requirements.txt
nonebot2[fastapi]
nonebot-adapter-onebot
redis
httpx
pyyaml
websockets
minio
Pillow
SQLAlchemy[asyncio]
aiomysql
cryptography
PicImageSearch
cloudscraper
pyquery